blueos_startup_update, commonwealth: extract cpu/host detection to commonwealth
